requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.7.0
django-extensions==4.1
django-vite==3.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.8.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
